Dividi em endpoints como pedia o exercício
